var-product-intelligence/pyproject.toml
[project]
name = "var-product-intelligence"
version = "1.0.0"
description = "Product comparison and recommendation platform for Value-Added Resellers"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy>=2.0.25",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # AI/LLM
    "anthropic>=0.18.0",

    # PDF Generation
    "reportlab>=4.0.0",

    # URL Fetching & HTML Parsing
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",

    # MCP Server
    "fastmcp>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
